cleaned up profile page edit widgets and added animation to statistics page
